requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.0.0

# For OpenAI tokenizer
tiktoken>=0.5.0

# For testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Formatting and linting
black>=23.0.0
flake8>=6.0.0
